requirements.txt
tensorflow-gpu==1.10
keras==2.1.5
matplotlib
h5py